Move store-related classes to quasar.run.store
