+++ run/src/test/scala/quasar/run/store/StoreCodecSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.run
+package quasar.run.store
 
 import slamdata.Predef._
